Move from types to dpnp_ndarray
